+++ src/main.rs
@@ -95,6 +95,7 @@ pub async fn inner() -> anyhow::Result<()> {
         Commands::Run(args) => commands::run::run(args.apply(config)).await,
         Commands::Validate(args) => commands::validate::validate(args.apply(config)).await,
         Commands::Dev(commands::DevCommands::Doc(args)) => commands::doc::doc(args).await,
+        Commands::Dev(commands::DevCommands::Lock(args)) => commands::lock::lock(args).await,
     }
 }