+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.14.0 - 07-09-2025
+
 ### Changed
 
 * Removed the `--config` option of `sprocket run`; the run command's